requirements_bigdata.txt
# 原有依赖
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
bcrypt==4.0.1
waitress==2.1.2
PyMySQL==1.1.0
cryptography==41.0.7

# 大数据技术栈依赖
# Hadoop生态系统
hdfs3==0.3.1
pyhive==0.7.0
thrift==0.16.0

# Spark相关
pyspark==3.4.1
findspark==2.0.1

# 数据处理
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# 机器学习
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
paddlepaddle==2.5.1

# 数据可视化
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
wordcloud==1.9.2
pyecharts==1.9.1

# 网络爬虫
scrapy==2.10.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2

# 数据库连接
pymongo==4.4.1
redis==4.6.0
sqlalchemy==2.0.19

# 工具库
hyperopt==0.2.7
joblib==1.3.2
tqdm==4.65.0
python-dotenv==1.0.0

# 报告导出功能
openpyxl==3.1.2
reportlab==4.0.4

# Web框架增强
Flask-SocketIO==5.3.6
gunicorn==21.2.0

# 日志和监控
loguru==0.7.0
prometheus-client==0.17.1
